requirements.txt
beautifulsoup4~=4.12.3
fastapi~=0.115.12
google-generativeai~=0.8.4
lxml~=5.3.2
python-dotenv~=1.1.0
requests~=2.32.3
uvicorn~=0.34.0
xmltodict~=0.14.2
certifi~=2025.1.31
protobuf~=5.29.4
schedule==1.2.0
Flask==2.3.3
Flask-CORS==4.0.0
gunicorn==21.2.0
python-dotenv~=1.1.0